Move food to individual file
