+++ exercises.html
@@ -10,10 +10,10 @@
       <div class="col s12 m3 manage-block">
         <h4>Manage Exercises</h4>
         <form class="exercises-form">
-          <label for="exercise-name">Name</label>
-          <input type="text" name="exercise-name" />
-          <label for="exercise-calories">Calories</label>
-          <input type="text" name="exercise-calories" />
+          <label for="name">Name</label>
+          <input type="text" name="name" />
+          <label for="calories">Calories</label>
+          <input type="text" name="calories" />
           <input type="submit" value="Add Exercise" class="btn" />
         </form>
       </div>
