+++ foods.html
@@ -10,10 +10,10 @@
       <div class="col s12 m3 manage-block">
         <h4>Manage Foods</h4>
         <form class="foods-form">
-          <label for="food-name">Name</label>
-          <input type="text" name="food-name"/>
-          <label for="food-calories">Calories</label>
-          <input type="text" name="food-calories"/>
+          <label for="name">Name</label>
+          <input type="text" name="name"/>
+          <label for="calories">Calories</label>
+          <input type="text" name="calories"/>
           <input type="submit" value="Add Food" class="btn" />
         </form>
       </div>
