+++ lib/index.js
@@ -1,6 +1,7 @@
 require("./css/materialize.css");
 require("./materialize.js");
 require("./css/main.css");
+var Food = require("./food")
 
 function addRow(name, calories, type) {
   $(type).append(
@@ -9,22 +10,16 @@ function addRow(name, calories, type) {
 }
 
 function addFood() {
-  var name = $('input[name="food-name"]').val();
-  var calories = $('input[name="food-calories"]').val();
+  var food = new Food($('input[name="name"]').val(),
+    $('input[name="calories"]').val());
   
-  storeFood(name, calories);
-  addRow(name, calories, '.food-table');
-}
-
-function storeFood(name, calories) {
-  var foodJSON = localStorage.getItem('foods');
-  if (foodJSON === null) {
-    foodJSON = '[]';
+  if (food.name === "") {
+    $('input[name="name"]')
+      .after('<p class="name-err">Please enter a food name</p>')
+  } else {
+    food.store();
+    addRow(name, calories, '.food-table');
   }
-  var foods = JSON.parse(foodJSON);
-  foods.push({name: name, calories: calories});
-  foodJSON = JSON.stringify(foods);
-  localStorage.setItem('foods', foodJSON);
 }
 
 function loadFoods() {
@@ -38,8 +33,8 @@ function loadFoods() {
 }
 
 function addExercise() {
-  var name = $('input[name="exercise-name"]').val();
-  var calories = $('input[name="exercise-calories"]').val();
+  var name = $('input[name="name"]').val();
+  var calories = $('input[name="calories"]').val();
   
   storeExercise(name, calories);
   addRow(name, calories, '.exercise-table');
